backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fintech-compliance-engine"
version = "0.1.0"
description = "AI-powered compliance automation for fintech code repositories"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.18",
    "pgvector>=0.2.4",
    "redis>=5.0.1",
    "rq>=1.16.1",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "tree-sitter>=0.21.0",
    "pygit2>=1.14.0",
    "azure-storage-blob>=12.19.0",
    "azure-ai-formrecognizer>=3.3.2",
    "openai>=1.10.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.1",
    "isort>=5.13.2",
    "ruff>=0.1.13",
    "mypy>=1.8.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=app --cov-report=term-missing"
